requirements.txt
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
dash==2.7.1
dash-bootstrap-components==1.2.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
db-dtypes==1.0.5
Flask==2.2.2
future==0.18.2
google-api-core==2.11.0
google-auth==2.15.0
google-auth-oauthlib==0.8.0
google-cloud-bigquery==3.4.1
google-cloud-bigquery-storage==2.16.2
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
grpcio==1.51.1
grpcio-status==1.51.1
idna==3.4
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.5
oauthlib==3.2.2
packaging==21.3
pandas==1.5.2
pandas-gbq==0.18.1
plotly==5.11.0
proto-plus==1.22.1
protobuf==4.21.11
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydata-google-auth==1.4.0
pyparsing==3.0.9
pyserial==3.5
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tenacity==8.1.0
tqdm==4.64.1
urllib3==1.26.13
Werkzeug==2.2.2
gunicorn==20.1.0
dash-daq==0.5.0
Pyrebase4==4.6.0
